chore(lab): lint hygiene and lab-path golangci exclusions
Fix US spelling in lab harnesses, restore scheduler fits call sites,
and exclude research-only scripts/bench-remote-server from noisy
gosec/unparam/errcheck while keeping production packages strict.
Also keep doltserver Commit failure rollback semantics.

+++ scripts/bench-remote-server/gateway/cluster_ack.go
@@ -406,7 +406,7 @@ func int64Value(value any) (int64, bool) {
 	case int:
 		return int64(typed), true
 	case uint64:
-		if typed > uint64(^uint64(0)>>1) {
+		if typed > (^uint64(0) >> 1) {
 			return 0, false
 		}
 		return int64(typed), true
